Add monitoring capabilities to wait for realityserver.
If the monitorFrequency option is given, and is > 0 then the result is an
EventEmitter. It will attempt to connect to RealityServer every
monitorFrequency ms and if the connection fails it will emit a 'disconnected'
event. When RealityServer is connectale again it will emit 'connected'.

Monitoring can be stopped by calling the 'stop' function on the result.